Improve tabbar architecture
