+++ Teacher Workout/Features/Discover/Coordinator/DiscoverCoordinator.swift
@@ -1,0 +1,15 @@
+import Foundation
+import UIKit
+
+final class DiscoverCoordinator: NSObject, Coordinator {
+    var childCoordinators: [Coordinator] = []
+    var navigationController: UINavigationController
+
+    override init() {
+        let viewController = DiscoverViewController()
+        viewController.tabBarItem.title = AppStrings.Menu.discover.localized()
+        viewController.tabBarItem.image = UIImage(systemName: "magnifyingglass")
+        navigationController = UINavigationController(rootViewController: viewController)
+        navigationController.setNavigationBarHidden(true, animated: false)
+    }
+}
