+++ Teacher Workout/Features/Home/Coordinator/HomeCoordinator.swift
@@ -1,0 +1,15 @@
+import Foundation
+import UIKit
+
+final class HomeCoordinator: NSObject, Coordinator {
+    var childCoordinators: [Coordinator] = []
+    var navigationController: UINavigationController
+
+    override init() {
+        let viewController = HomeViewController()
+        viewController.tabBarItem.title = AppStrings.Menu.home.localized()
+        viewController.tabBarItem.image = UIImage(systemName: "house")
+        navigationController = UINavigationController(rootViewController: viewController)
+        navigationController.setNavigationBarHidden(true, animated: false)
+    }
+}
