+++ Teacher Workout/Features/Profile/Coordinator/ProfileCoordinator.swift
@@ -1,0 +1,15 @@
+import Foundation
+import UIKit
+
+final class ProfileCoordinator: NSObject, Coordinator {
+    var childCoordinators: [Coordinator] = []
+    var navigationController: UINavigationController
+
+    override init() {
+        let viewController = ProfileViewController()
+        viewController.tabBarItem.title = AppStrings.Menu.profile.localized()
+        viewController.tabBarItem.image = UIImage(systemName: "person.circle")
+        navigationController = UINavigationController(rootViewController: viewController)
+        navigationController.setNavigationBarHidden(true, animated: false)
+    }
+}
